+++ src/plugins/proxy/assets/proxy-stop.sh
@@ -3,6 +3,6 @@
 APPNAME=<%= appName %>
 
 sudo docker rm -f $APPNAME || :
-docker network disconnect bridge -f $APPNAME || :
+sudo docker network disconnect bridge -f $APPNAME || :
 
 sudo docker rm -f $APPNAME-letsencrypt || :